move table partial into own file, lint format imports
